sih-ML/requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.8.0
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==3.1.0
boto==2.49.0
boto3==1.10.44
botocore==1.13.44
Brlapi==0.6.6
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorlover==0.3.0
command-not-found==0.3
ConfigArgParse==0.14.0
constantly==15.1.0
contractions==0.0.24
cryptography==2.1.4
cssselect==1.1.0
cufflinks==0.16
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.0
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
dlib==19.16.0
docutils==0.15.2
editdistance==0.3.1
entrypoints==0.3
evdev==0.7.0
Flask==1.0
future==0.17.1
gensim==3.8.1
h5py==2.10.0
html5lib==0.999999999
httplib2==0.9.2
hyperlink==19.0.0
idna==2.7
incremental==17.5.0
ipykernel==5.1.1
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.11
jsonschema==3.0.2
jupyter-client==5.3.1
jupyter-core==4.5.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==3.1.1
mistune==0.8.4
mlxtend==0.17.0
nbconvert==5.6.0
nbformat==4.4.0
netifaces==0.10.4
nltk==3.4.5
notebook==6.0.1
npm==0.1.1
numpy==1.16.4
oauth==1.0.1
olefile==0.45.1
opencv-python==4.1.0.25
optional-django==0.1.0
pandas==0.25.0
pandocfilters==1.4.2
parsel==1.5.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
plotly==3.10.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
Protego==0.1.16
protobuf==3.0.0
ptyprocess==0.6.0
pyahocorasick==1.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyDispatcher==2.0.5
pyfpgrowth==1.0
Pygments==2.4.2
pygobject==3.26.1
PyHamcrest==1.9.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==19.1.0
pyparsing==2.4.2
pyRFC3339==1.0
pyrsistent==0.15.4
python-apt==1.6.4
python-dateutil==2.8.0
python-debian==0.1.32
pytz==2019.2
pyudev==0.21.0
pyxdg==0.25
PyYAML==3.12
pyzmq==18.0.2
queuelib==1.5.0
reportlab==3.4.0
requests==2.20.0
requests-unixsocket==0.1.5
retrying==1.3.3
s3transfer==0.2.1
scikit-learn==0.19.1
scipy==0.19.0
Scrapy==1.8.0
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
simplejson==3.13.2
six==1.12.0
sklearn==0.0
smart-open==1.9.0
system-service==0.3
systemd-python==234
terminado==0.8.2
testpath==0.4.2
textsearch==0.0.17
torch==1.1.0
torchtext==0.4.0
torchvision==0.3.0
tornado==6.0.3
tqdm==4.30.0
traitlets==4.3.2
Twisted==19.10.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.1.1
urllib3==1.24.3
usb-creator==0.3.3
virtualenv==16.7.9
w3lib==1.21.0
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
xkit==0.0.0
zope.interface==4.3.2
